integrations-examples/pipecat-bot/pyproject.toml
[project]
name = "pipecat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pipecat-ai[webrtc,gladia,runner]>=0.0.80",
    "python-dotenv>=1.1.1",
]
